Add ignored attributes when deserial. from XML
The serializer interpreted the ID attribute from the XML file as one
attribute to be imported, which wasn't the case. This attempt triggered
a Symfony\Component\Serializer\Exception\NotNormalizableValueException,
as actually the "id" attribute of the entity expects a "int" value, and
the serializer was deserializing the "ID" as a string.