Simplify Milestone entity and remove components
As milestones in GanttProject can't be longer than one day, there is
no need of having an "start date" and a "finish date". Just having
a "date" attribute in the entity is enough.

Also, the symfony/serializer component has been removed as maybe using
it could be considered as overkill, because it was only used to
deserialize a couple of attributes.