Fix: modify fk constraints of user and ganttproject
The ganttproject table doesn't need a "user_id" field, as the model of
the application sets the owning side of the relation to the user entity.

The user table should not delete the entries when the associated
ganttproject is deleted. Instead, it should set the entries'
ganttproject field to null.