Refactor: make XmlUtils use Doctrine
In this change XmlUtils uses directly Doctrine, and therefore the
function that deals with storing the tasks in the database needed a
GanttProject to associate the tasks with.

The tests have been modified accordingly, and Doctrine's data fixtures
are used for data loading into the database.

As Doctrine inserts empty strings instead of nulls when inserting a task
with no name, instead of modifying the setName of the Task entity, the results
that the test expects have been changed from null to empty string.