Feature: modify a task duration and its dependants' date
These changes are stored both in the database and in the XML.

The Task entity's delayDate operation was changed to introduce an object
cloning, as otherwise Doctrine is not able to detect that changes were
made to the DateTime object.

Finally the GanttProjectDataLoader was modified to change the fixture's
name to something that was going to be checked afterwards in the tests.
Until now that name was ignored, and therefore it doesn't affect any
other test.