Refactor: make structure.sql MySQL compliant
MySQL doesn't seem to support `IF NOT EXISTS` when modifying a table's
column, while MariaDB —the database used in the development environment—
does. Instead of creating two versions of the SQL file —one for MySQL
and another one for MariaDB— I have decided to go simple and make the
file MySQL compliant for the sake of mantainability.