Chore: improve README by replacing numbers by references in links
When editing the README one had to check at the bottom which link were
the numbers referring to. By introducing references this problem is
solved, plus the references can be in no particular order instead of
having to reorder the whole thing whenever a new link was introduced in
between.